requirements.txt
Flask == 1.1.1
Flask-RESTful == 0.3.8
flask-restplus == 0.13.0
Flask-SQLAlchemy == 2.4.1
Jinja2 == 2.11.1
SQLAlchemy == 1.3.15
requests == 2.23.0
flask_marshmallow == 0.11.0
flask_script == 2.0.6
flask_migrate == 2.5.3
marshmallow-sqlalchemy == 0.22.3
mysqlclient == 1.4.6
pyjwt == 1.7.1
Werkzeug == 0.16.1
pytest-flask == 1.0.0


